FE: Move `SignupPersonalDataPage` file into separated folder
